refactor: UI 컴포넌트 파일 구조 정리 및 네이밍 통일
- 중복된 Button.tsx, button-variants.ts 파일 삭제
- kebab-case 파일명을 PascalCase로 변경 (breadcrumb, scroll-area, toast, toaster)
- 직접 경로 import를 barrel export 활용하도록 변경
- 빌드 에러 해결 및 네이밍 컨벤션 준수
