+++ src/hooks/useToast.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 
-import type { ToastActionElement, ToastProps } from '@/components/ui/toast';
+import type { ToastActionElement, ToastProps } from '@/components/ui/Toast';
 
 const TOAST_LIMIT = 5;
 const TOAST_REMOVE_DELAY = 5000;
